requirements.txt
PyGithub
bandit
safety
loguru
simplejson